+++ merkletree/merkletreekv.go
@@ -1,9 +1,9 @@
 package merkletree
 
 import (
-	"github.com/coniks-sys/libmerkleprefixtree-go/crypto"
-	"github.com/coniks-sys/libmerkleprefixtree-go/internal"
-	"github.com/coniks-sys/libmerkleprefixtree-go/kv"
+	"github.com/coniks-sys/coniks-go/crypto"
+	"github.com/coniks-sys/coniks-go/kv"
+	"github.com/coniks-sys/coniks-go/utils"
 )
 
 func NewMerkleTreeFromKV(db kv.DB, epoch uint64) (*MerkleTree, error) {
