+++ merkletree/node_test.go
