+++ merkletree/strkv.go
@@ -3,9 +3,9 @@ package merkletree
 import (
 	"encoding/binary"
 
-	"github.com/coniks-sys/libmerkleprefixtree-go/crypto"
-	"github.com/coniks-sys/libmerkleprefixtree-go/internal"
-	"github.com/coniks-sys/libmerkleprefixtree-go/kv"
+	"github.com/coniks-sys/coniks-go/crypto"
+	"github.com/coniks-sys/coniks-go/kv"
+	"github.com/coniks-sys/coniks-go/utils"
 )
 
 func serializeSTRKVKey(epoch uint64) []byte {
